Add check for placeholder image

+++ scripts/fpp_install.sh
@@ -26,6 +26,11 @@ else
 	sudo chown fpp /home/fpp/media/config/plugin.fpp-jukebox.json
 fi
 
-echo "Copy placeholder image to images folder"
-sudo cp /home/fpp/media/plugins/fpp-julebox/img/placeholder.jpg /home/fpp/media/images/placeholder.jpg
-sudo chown fpp /home/fpp/media/images/placeholder.jpg
+
+placeholder_image =/home/fpp/media/images/placeholder.jpg
+if [ -s "$placeholder_image" ]
+	echo "Placehoolder image found no"
+else
+	echo "Placehoolder image not found, Copy placeholder image to images folder"
+	sudo cp /home/fpp/media/plugins/fpp-julebox/img/placeholder.jpg /home/fpp/media/images/placeholder.jpg
+	sudo chown fpp /home/fpp/media/images/placeholder.jpg
